procurement-pipeline-project/requirements.txt
pandas==2.1.4
faker==22.0.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
hdfs==2.7.0
sqlalchemy==2.0.23
requests>=2.31.0
numpy==1.26.4
